Changed some variables, properties, and parameters of type int to sbyte to reduce memory usage

+++ NQueen.Common/Interface/ISimulationResults.cs
@@ -4,9 +4,12 @@ namespace NQueen.Common.Interface
 {
     public interface ISimulationResults
     {
-        int BoardSize { get; set; }
+        sbyte BoardSize { get; set; }
+
         IEnumerable<Solution> Solutions { get; set; }
+
         int NoOfSolutions { get; set; }
+
         double ElapsedTimeInSec { get; set; }
     }
 }
